backend/requirements.txt
fastapi
uvicorn[standard]
playwright
playwright-stealth
requests
supabase
python-dotenv
